requirements.txt
requests
python-dotenv
nvidia-ml-py
